Make Element, Colour, Float3 classes immutable
All relevant handlers and editors have been modified to suit.

This makes it a lot easier to deal with passed references
The downside is a slight overhead in construction and garbage collection
This shouldn't happen at render time, so it doesn't matter too much
A slight delay might occur in the UI, for instance when saving an element
with many vertices, or transforming a large composite shape.